+++ c_src/repository.c
@@ -1,4 +1,5 @@
 #include "repository.h"
+#include "credential.h"
 #include "object.h"
 #include "odb.h"
 #include "oid.h"
@@ -6,6 +7,7 @@
 #include "index.h"
 #include "ex_git_engine.h"
 #include <string.h>
+#include <stdlib.h>
 #include <git2.h>
 
 void git_engine_repository_free(ErlNifEnv *env, void *cd)
@@ -243,6 +245,8 @@ git_engine_repository_clone(ErlNifEnv *env, int argc, const ERL_NIF_TERM argv[])
 	ERL_NIF_TERM term_repo;
 	git_clone_options opts = GIT_CLONE_OPTIONS_INIT;
 	git_strarray headers = { NULL, 0 };
+	git_engine_credential_res *cred_res = NULL;
+	git_engine_credential_payload cred_payload;
 
 	if (!enif_inspect_binary(env, argv[0], &url))
 		return enif_make_badarg(env);
@@ -259,11 +263,29 @@ git_engine_repository_clone(ErlNifEnv *env, int argc, const ERL_NIF_TERM argv[])
 	bare = !enif_compare(argv[2], atoms.true);
 	opts.bare = bare ? 1 : 0;
 
+	/* argv[3]: static headers (non-auth) */
 	headers = git_strarray_from_list(env, argv[3]);
 	opts.fetch_opts.custom_headers = headers;
 
+	/* argv[4]: runner PID for credential callback, or nil/false to skip */
+	if (enif_get_local_pid(env, argv[4], &cred_payload.runner_pid)) {
+		cred_res = enif_alloc_resource(git_engine_credential_type,
+		                               sizeof(git_engine_credential_res));
+		memset(cred_res, 0, sizeof(git_engine_credential_res));
+		cred_res->mtx = enif_mutex_create((char *)"credential_mtx");
+		cred_res->cond = enif_cond_create((char *)"credential_cond");
+
+		cred_payload.env      = env;
+		cred_payload.res      = cred_res;
+		cred_payload.res_term = enif_make_resource(env, cred_res);
+
+		opts.fetch_opts.callbacks.credentials = git_engine_credential_acquire_cb;
+		opts.fetch_opts.callbacks.payload     = &cred_payload;
+	}
+
 	error = git_clone(&repo, (char *)url.data, (char *)local_path.data, &opts);
 	git_strarray_free(&headers);
+	if (cred_res) enif_release_resource(cred_res);
 	if (error < 0)
 		return git_engine_error_struct(env, error);